% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RPS.R
\name{plot_RPS}
\alias{plot_RPS}
\title{Creating of whole plots by selection}
\usage{
plot_RPS(sampH0 = sampH0, sampH1 = sampH1, sampEmp = sampEmp, select = 6)
}
\arguments{
\item{sampH0}{The result of samplesH0 function}

\item{sampH1}{The result of samplesH1 function}

\item{sampEmp}{The result of samplesEmp function}

\item{select}{The selection of plot options}
}
\value{
Plot summary of ResProStr. Based on the selection, different plots
are generated;
select = 1: Distribution of t-values for given H0 and H1
select = 2: Proportion of LR as a function of criteria
select = 3: Distribution of t-values for given H0 and H1 adapted to d observed
select = 4: Proportion of LR as a function of selected criteria.
select = 5: Distributions of t-values for given H1(=t empirical) und H0
select = 6: Multiple plots for the case of the likelihood ratios prior to data collection
}
\description{
Creating of whole plots by selection
}
\examples{
\dontrun{
plot_RPS(sampH0 = sampH0, sampH1 = sampH1, sampEmp = sampEmp,
select = 6)
}
}
