% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples30.R
\name{samples30}
\alias{samples30}
\title{Creating samples from H1, given alpha, effect size (d) and power}
\usage{
samples30(Nsample = 100, alpha, effectSize, pow, samp)
}
\arguments{
\item{Nsample}{It can be fixed or given by the user}

\item{alpha}{level of significance (α = probability of type I error)}

\item{effectSize}{the hypothesized effect size}

\item{pow}{the desired power (1−β)}

\item{samp}{The output of samplesH1 function, it is necessary for Nest here}
}
\value{
Simulated likelihood ratios L1x, L0x and Lplus0x
}
\description{
Creating samples from H1, given alpha, effect size (d) and power
}
\examples{
\dontrun{
samples30(Nsample = 100, alpha = 0.05, effectSize = 0.2, pow = 0.95, samp = samplesH1)
}

}
