% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesEmp.R
\name{samplesEmp}
\alias{samplesEmp}
\title{Simulate likelihood ratios post-hoc}
\usage{
samplesEmp(Nsample = 100, N, Temp, alphaEmp)
}
\arguments{
\item{Nsample}{It can be fixed or given by the user}

\item{N}{The sample size N per group}

\item{Temp}{Empirical t-value}

\item{alphaEmp}{Empirical level of significance}
}
\value{
Simulate likelihood ratios with
demp: obtained empirical effectsize
pow: the desired power (1−β)
Tval: t value
p: Calculated PCDF values D|L(d=0|x) ?
}
\description{
Simulate likelihood ratios post-hoc
}
\examples{
\dontrun{
samplesEmp(Nsample = 100, N = 30, Temp = 3, alphaEmp = 0.05)
}

}
