% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesEmp0.R
\name{samplesEmp0}
\alias{samplesEmp0}
\title{Simulate likelihood ratios post-hoc under H0 ?}
\usage{
samplesEmp0(Nsample, N, Temp, alphaEmp)
}
\arguments{
\item{Nsample}{It can be fixed or given by the user}

\item{N}{The sample size N per group}

\item{Temp}{Empirical t-value}

\item{alphaEmp}{Empirical level of significance}
}
\value{
Simulate likelihood ratios with
p: Calculated PCDF values D|L(d=0|x) ?
Tval: t value
ncp: non-centrality-parameter of the t-distribution
demp: obtained empirical effectsize
}
\description{
Simulate likelihood ratios post-hoc under H0 ?
}
\examples{
\dontrun{
samplesEmp0(Nsample = 100, N = 30, Temp = 3, alphaEmp = 0.05)
}

}
