% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesH0.R
\name{samplesH0}
\alias{samplesH0}
\title{Sampling from a distribution of test results given H0}
\usage{
samplesH0(Nsample = 100, alpha, effectSize, pow)
}
\arguments{
\item{Nsample}{It can be fixed or given by the user}

\item{alpha}{level of significance (α = probability of type I error)}

\item{effectSize}{the hypothesized effect size}

\item{pow}{the desired power (1−β)}
}
\value{
Simulated likelihood ratios with
p: Calculated PCDF values D|L(d=0|x) ?
Tval: t value
Nest: estimated group size
ncp: non-centrality-parameter of the t-distribution representing H1
demps: obtained empirical effectsize
}
\description{
Sampling from a distribution of test results given H0
}
\examples{
\dontrun{
samplesH0(Nsample = 100, alpha = 0.05, effectSize = 0.2, pow = 0.95)
}

}
