% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumLR.R
\name{sumLR}
\alias{sumLR}
\title{Samples from different H relative to H1, given alpha, and Nest d}
\usage{
sumLR(Nsample, alpha, effectSize, pow, samp)
}
\arguments{
\item{Nsample}{It can be fixed or given by the user}

\item{alpha}{level of significance (α = probability of type I error)}

\item{effectSize}{the hypothesized effect size}

\item{pow}{the desired power (1−β)}

\item{samp}{The output of samplesH1 function}
}
\value{
The proportions of simulated t-values for the RSP methodology
Here, the Wald-criterion (1−β)/α is applied for interpreting the obtained
likelihood ratios
}
\description{
Samples from different H relative to H1, given alpha, and Nest d
}
\examples{
\dontrun{
sumLR(Nsample = 100, alpha = 0.05, effectSize = 0.2, pow = 0.95, samp = samplesH1)
}

}
