% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_RPS.R
\name{table_RPS}
\alias{table_RPS}
\title{Creating table outputs by selection}
\usage{
table_RPS(
  sampH0 = sampH0,
  sampH1 = sampH1,
  sampEmp = sampEmp,
  sampEmp0 = sampEmp0,
  samp30 = samp30,
  select = 1
)
}
\arguments{
\item{sampH0}{The result of samplesH0 function}

\item{sampH1}{The result of samplesH1 function}

\item{sampEmp}{The result of samplesEmp function}

\item{sampEmp0}{The result of samplesEmp0 function}

\item{samp30}{The result of samples30 function}

\item{select}{The selection of table options (default is 1 for Proportion of correct positive results
(prior to data collection))}
}
\value{
Table wrap-up function of RPS. Based on the selection, different tables
are generated;
select = 1: Proportion of correct positive results (prior to data collection)
select = 2: Proportion of false positive results (prior to data collection)
select = 3: Proportion of correct positive results (post-hoc)
select = 4: Proportion of false positive results (post-hoc)
}
\description{
Creating table outputs by selection
}
\examples{
\dontrun{
table_RPS(sampH0 = sampH0, sampH1 = sampH1, sampEmp = sampEmp, sampEmp0 = sampEmp0,
samp30 = samp30, select = 1)
}
}
